Backend/pyproject.toml
[project]
name = "nyaayveer-backend"
version = "0.1.0"
description = "NyaayVeer is a mobile-first legal assistant designed to make criminal law information more accessible to everyday users."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-caching>=2.3.1",
    "flask-cors>=5.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "groq>=0.23.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "xgboost>=3.0.0",
]
